+++ rectangleFunction.cpp
@@ -5,6 +5,6 @@ int rectanglePerimeter(int firstEdge, int secondEdge){
 int rectangleArea(int firstEdge, int secondEdge){
 	return firstEdge*secondEdge;
 }
-int rectangleDiagonal(int firstEdge, int secondEdge){
+float rectangleDiagonal(int firstEdge, int secondEdge){
 	return  sqrt(firstEdge*firstEdge+secondEdge*secondEdge);
 }
